refactoring for packaging + gitignore update
