+++ setup.cfg
@@ -1,0 +1,24 @@
+[metadata]
+name = DeepRL
+version = 0.0.1
+author = Yann Berthelot
+author_email = yannberthelot1@gmail.com
+description = A small package for Deep RL agents
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/YannBerthelot/DeepRL
+# project_urls =
+#     Bug Tracker = https://github.com/pypa/sampleproject/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.8
+
+[options.packages.find]
+where = src
