+++ src/DeepRL/main.py
@@ -1,0 +1,4 @@
+from pymgrid_exp.pymgrid_experiment import pymgrid_experiment
+
+if __name__ == "__main__":
+    pymgrid_experiment()
