+++ src/DeepRL/network/utils.py
@@ -1,6 +1,7 @@
 import torch
 import torch.nn as nn
 from typing import List
+import warnings
 
 # helper function to convert numpy arrays to tensors
 def t(x):
@@ -105,3 +106,34 @@ def compute_KL_divergence(
     else:
         KL_divergence = 0
     return KL_divergence
+
+
+def get_device(device_name: str) -> torch.DeviceObjType:
+    """
+    Chose the right device for PyTorch. If no GPU is available, it will use CPU.
+
+    Args:
+        device_name (str): The device to use between "GPU" and "CPU"
+
+    Returns:
+        torch.DeviceObjType: The Torch.Device to use
+    """
+    if device_name == "GPU":
+        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+        if not torch.cuda.is_available():
+            warnings.warn("GPU not available, switching to CPU", UserWarning)
+    else:
+        device = torch.device("cpu")
+
+    return device
+
+
+class LinearSchedule:
+    def __init__(self, start, end, t_max) -> None:
+        self.start = start
+        self.end = end
+        self.t_max = t_max
+        self.step = (start - end) / t_max
+
+    def transform(self, t):
+        return self.start - self.step * t
