+++ src/DeepRL/utils/buffer.py
@@ -1,5 +1,4 @@
 import numpy as np
-import torch
 
 
 class RolloutBuffer:
